% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_probability_of_freedom.R
\name{compute_probability_of_freedom}
\alias{compute_probability_of_freedom}
\title{Compute probability of disease freedom given samples taken per timestep}
\usage{
compute_probability_of_freedom(
  n,
  phi_prior,
  alpha_prior,
  beta_prior,
  alpha_intro,
  beta_intro,
  p_intro,
  growth_rate,
  rho,
  pi = 0,
  delta_t = 1,
  pi_seq = 1000
)
}
\arguments{
\item{n}{number of samples taken per timestep}

\item{phi_prior}{prior probability of freedom}

\item{alpha_prior}{beta distribution alpha parameter for prior prevalence distribution}

\item{beta_prior}{beta distribution beta parameter for prior prevalence distribution}

\item{alpha_intro}{beta distribution alpha shape parameter}

\item{beta_intro}{beta distribution beta shape parameter}

\item{p_intro}{probability of introduction; if provided as a vector, the
initial entry corresponds to the probability of introduction
at time step 2.}

\item{growth_rate}{exponential growth rate}

\item{rho}{unit test sensitivity}

\item{pi}{design threshold}

\item{delta_t}{test time duration}

\item{pi_seq}{discretization granularity}
}
\value{
"diseasefree" structure including sample sizes and prior/posterior distributions
}
\description{
Compute probability of disease freedom given samples taken per timestep
}
\examples{
u <- compute_probability_of_freedom(c(21,4,4,4,4), 0.5, 1, 1, 1, 1, 0.04, 0.01, 0.9, 0.0)
summary(u)
plot(u)
}
