% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sample_size.R
\name{compute_sample_size}
\alias{compute_sample_size}
\title{Compute the sample size requirements to meet a specified design
threshold at a specified confidence.}
\usage{
compute_sample_size(
  phi_prior,
  alpha,
  beta,
  p_intro,
  growth_rate,
  rho,
  pi,
  dconf,
  delta_t = 1,
  n_steps = 1,
  method = "restore",
  pi_seq = 1000
)
}
\arguments{
\item{phi_prior}{prior probability of freedom}

\item{alpha}{beta distribution alpha shape parameter}

\item{beta}{beta distribution beta shape parameter}

\item{p_intro}{probability of introduction}

\item{growth_rate}{exponential growth rate}

\item{rho}{unit test sensitivity}

\item{pi}{design threshold}

\item{dconf}{desired confidence}

\item{delta_t}{test time duration}

\item{n_steps}{number of time units to step}

\item{method}{"restore" or "maintain"}

\item{pi_seq}{discretization granularity}
}
\value{
"diseasefree" structure including sample sizes and prior/posterior distributions
}
\description{
Compute the sample size requirements to meet a specified design
threshold at a specified confidence.
}
\examples{
u <- compute_sample_size(0.5, 1, 1, 0.04, 0.01, 0.9, 0, 0.95, n_steps=5)
summary(u)
plot(u)
}
